// front/src/client/types.gen.ts
export type ExerciseModel = {
    muscle_group_id: number;
    muscle_id: number;
    exercise_id: number;
    exercise_name: string;
    sets?: Array<back__models__exercise__ExerciseModel__setDetail>;
};

export type ExercisedetailsModel = {
    user_id: number;
    exercise_id: number;
    exercise_details_id: number;
    performed_date?: string;
    sets: Array<back__models__exerciseDetails__ExercisedetailsModel__setDetail>;
};

export type FetchAll = {
    muscle_name: string;
    muscle_group_name: string;
    exercise_name: string;
    performed_date: string;
    user_id: number;
    exercise_id: number;
    exercise_details_id: number;
};

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type LoginModel = {
    name: string;
    password: string;
};

export type LoginResponse = {
    success: boolean;
};

export type MuscleGroupModel = {
    muscle_group_id: number;
    muscle_group_name: string;
};

export type MuscleModel = {
    muscle_group_id: number;
    muscle_id: number;
    muscle_name: string;
};

export type UserModel = {
    user_id: number;
    name: string;
    password: string;
    gender: string;
    age: number;
    height: number;
    weight: number;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
};

export type back__models__exerciseDetails__ExercisedetailsModel__setDetail = {
    set_number: number;
    weight: number;
    reps: number;
};

export type back__models__exercise__ExerciseModel__setDetail = {
    set_number: number;
    weight: number;
    reps: number;
};

export type ExerciseDetailsFetchSetDetailsData = {
    query: {
        exercise_details_id: number;
        exercise_id: number;
        user_id: number;
    };
};

export type ExerciseDetailsFetchSetDetailsResponse = Array<back__models__exerciseDetails__ExercisedetailsModel__setDetail>;

export type ExerciseDetailsFetchSetDetailsError = unknown | HTTPValidationError;

export type ExerciseDetailsFetchAllData = {
    query?: {
        exercise_details_id?: number;
        exercise_id?: number;
        user_id?: number;
    };
};

export type ExerciseDetailsFetchAllResponse = Array<FetchAll>;

export type ExerciseDetailsFetchAllError = unknown | HTTPValidationError;

export type ExerciseDetailsInsertData = {
    body: ExercisedetailsModel;
};

export type ExerciseDetailsInsertResponse = number;

export type ExerciseDetailsInsertError = unknown | HTTPValidationError;

export type MuscleGroupsFetchAllData = {
    query?: {
        muscle_group_id?: number;
    };
};

export type MuscleGroupsFetchAllResponse = Array<MuscleGroupModel>;
// ...
export type MusclesFetchAllResponse = Array<MuscleModel>;

export type MusclesFetchAllError = unknown | HTTPValidationError;

export type ExercisesFetchAllData = {
    query?: {
        exercise_id?: number;
        muscle_group_id?: number;
        muscle_id?: number;
    };
};

export type ExercisesFetchAllResponse = Array<ExerciseModel>;

export type ExercisesFetchAllError = unknown | HTTPValidationError;

export type ExercisesFetchNextSetData = {
    query: {
        exercise_id: number;
        user_id: number;
    };
};

export type ExercisesFetchNextSetResponse = Array<back__models__exercise__ExerciseModel__setDetail>;